Misty/requirements.txt
openai
requests
SpeechRecognition
pyaudio
python-dotenv
